+++ lib/templatechecker/class.TemplateChecker.php
@@ -17,12 +17,12 @@ class TemplateChecker {
 	 * Array containing parameter information for custom widgeds
 	 * @var array
 	 */
-	private $customWidgets = array();
+	private $widgets = array();
 	
 	private $ignore_html_error_code = array(
 		68 => array(), //error parsing attribute name 
 		800 => array(), // Missplaced ***
-		801 => array('maxline' => 10),
+		//801 => array('maxline' => 10),
 	);
 
 	/**
@@ -45,26 +45,8 @@ class TemplateChecker {
 	 */
 	public static function run($fileName, MessageCollection $messages) {
 		$checker = new TemplateChecker($fileName, $messages);
-		$checker->injectCustomWidgets();
 		$checker->performTests();
 		return $checker;
-	}
-
-	/**
-	 * Try to read page-specific file containing additional widget information and merge
-	 * this information into the widget_params structure
-	 */
-	private function injectCustomWidgets() {
-		$file = const_path . 'pages/' . config_pages . '/templatechecker.customwidgets.php';
-		if (!is_file($file))
-			return;
-
-		try {
-			@require_once $file;
-			$this->customWidgets = $customWidgets;
-		} catch (Exception $ex) {
-			$this->messages->addError('CUSTOM WIDGETS', 'reading custom widget information failed: ' + $ex->getMessage());
-		}
 	}
 
 	/**
@@ -75,6 +57,7 @@ class TemplateChecker {
 	public function __construct($fileName, MessageCollection $messages) {
 		$this->messages = $messages;
 		$this->fileName = $fileName;
+		$this->widgets = twig_docu();
 	}
 
 	/**
@@ -103,7 +86,7 @@ class TemplateChecker {
 		// HTML FILE
 		$ext = strtolower(pathinfo($absFile, PATHINFO_EXTENSION));
 		if ($ext != 'html') {
-			$this->messages->addInfo('FILE CHECK', 'Not a html file: ' . $this->file);
+			$this->messages->addInfo('FILE CHECK', 'Not an html file: ' . $this->file);
 			return;
 		}
 
@@ -125,7 +108,7 @@ class TemplateChecker {
 		$errors = libxml_get_errors();
 		foreach ($errors as $error) {
 			if (array_key_exists($error->code, $this->ignore_html_error_code)) {
-				$condintions = $this->ignore_html_error_code[$error->code];
+				$conditions = $this->ignore_html_error_code[$error->code];
 				$ignore = true;
 				if ($conditions['maxline'] && $error->line > $conditions['maxline'])
 					$ignore = false;
@@ -167,7 +150,8 @@ class TemplateChecker {
 					return;
 			}
 		} else if ($node->nodeType == XML_TEXT_NODE) {
-			$value = trim($node->textContent);
+			// do not check code wrapped by {% verbatim %} or in twig comment /**...*/
+			$value = preg_replace('/{%\s*verbatim\s*%}.*?{%\s*endverbatim\s*%}|\/\*\*.*?\*\//', '', trim($node->textContent));
 			if (preg_match_all("/{{(.*?)}}/", $value, $macros)) {
 				for ($i = 0; $i < count($macros[0]); $i++)
 					$this->checkWidget($node, trim($macros[1][$i]));
@@ -221,20 +205,26 @@ class TemplateChecker {
 	 * @param \DOMElement $node currently checked node
 	 * @param string $macro widget (including parameters)	 
 	 */
-	private function checkWidget($node, $macro) {
+	public function checkWidget($node, $macro) {
 		$widget = Widget::create($node, $macro, $this->messages);
 		if ($widget == NULL)
 			return;
 
 		$widgetConfig = $this->getWidgetConfig($widget->getName());
-		if ($widgetConfig === NULL) {
-			$this->messages->addWarning('WIDGET PARAM CHECK', 'Unknown Widget found. Check manually!', $widget->getLineNumber(), $widget->getMacro(), $widget->getMessageData());
+
+		if ($widgetConfig !== NULL && array_key_exists('deprecated', $widgetConfig)) {
+			$this->messages->addWarning('WIDGET DEPRECATION CHECK', 'Deprecated widget', $widget->getLineNumber(), $widget->getMacro(), $widget->getMessageData());
+		}
+
+		if ($widgetConfig === NULL || !array_key_exists('param', $widgetConfig)) {
+			$this->messages->addWarning('WIDGET PARAM CHECK', 'Unknown widget found. Check manually!', $widget->getLineNumber(), $widget->getMacro(), $widget->getMessageData());
 			return;
 		}
 
 		// check all parameters of widget
-		foreach ($widgetConfig as $paramIndex => $paramConfig)
-			WidgetParameterChecker::performChecks($widget, $paramIndex, $paramConfig, $this->messages);
+		$paramConfig = array_values($widgetConfig['param']);
+		foreach ($paramConfig as $paramIndex => $paramConfig)
+			WidgetParameterChecker::performChecks($widget, $paramIndex, $paramConfig, $this->messages, $this);
 	}
 
 	/**
@@ -243,10 +233,8 @@ class TemplateChecker {
 	 * @return array parameter config for widget or NULL if widget is unknown
 	 */
 	private function getWidgetConfig($name) {
-		if (array_key_exists($name, $this->customWidgets)) {
-			return $this->customWidgets[$name];
-		} else if (array_key_exists($name, Config::SmartvisuDefaultWidgets)) {
-			return Config::SmartvisuDefaultWidgets[$name];
+		if (array_key_exists($name, $this->widgets)) {
+			return $this->widgets[$name];
 		} else {
 			return NULL;
 		}
@@ -272,8 +260,7 @@ class TemplateChecker {
 
 		// replace {{ icon0 }} and {{ icon1 }}
 		if (preg_match_all("/{{(.*?)}}/", $file, $match)) {
-			for ($i = 0; $i < count($match[0
-			]); $i++) {
+			for ($i = 0; $i < count($match[0]); $i++) {
 				switch (strtolower(trim($match[1] [$i]))) {
 					case 'icon0':
 						$file = str_replace($match[0][$i], $settings->getIcon0(), $file);
@@ -288,12 +275,30 @@ class TemplateChecker {
 			}
 		}
 
-		// add const_path if $file is relative
-		if (substr($file, 0, 1) != '/')
-			$file = const_path . $file;
+		// check different locations and variants if file has no path set
+    if(strpos($file, '/') === false) {
+			if(is_file(const_path . 'dropins/' . $settings->getIcon0() . $file))
+				return self::FILE_EXISTING;
+			if(is_file(const_path . 'dropins/' . $settings->getIcon0() . $file . '.svg'))
+				return self::FILE_EXISTING;
+			if(is_file(const_path . 'dropins/' . $settings->getIcon0() . 'jquery_' . $file . '.svg'))
+				return self::FILE_EXISTING;
 
-		$file = substr($file, 0, 1) == '/' ? $file : const_path . $file;
-		return is_file($file) ? self::FILE_EXISTING : self::FILE_MISSING;
+			if(is_file(const_path . $settings->getIcon0() . $file))
+				return self::FILE_EXISTING;
+			if(is_file(const_path . $settings->getIcon0() . $file . '.svg'))
+				return self::FILE_EXISTING;
+			if(is_file(const_path . $settings->getIcon0() . 'jquery_' . $file . '.svg'))
+				return self::FILE_EXISTING;
+
+			return self::FILE_MISSING;
+		}
+		else {
+			// add const_path if $file is relative
+			if (substr($file, 0, 1) != '/')
+				$file = const_path . $file;
+			return is_file($file) ? self::FILE_EXISTING : self::FILE_MISSING;
+		}
 	}
 
 }